src/backend/fastapi_app/requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
easyocr==1.7.2
elastic-transport==8.17.1
elasticsearch==8.14.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.3.2
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.6
joblib==1.5.0
lazy_loader==0.4
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
propcache==0.3.2
prov==2.0.1
puremagic==1.29
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-bidi==0.6.6
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyvi==0.1.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sklearn-crfsuite==0.5.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tika==3.1.0
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.1
